+++ features/home/navigation/src/main/java/com/tokopedia/navigation/presentation/customview/LottieBottomNavbar.kt
@@ -422,8 +422,8 @@ data class BottomMenu(val id: Int,
                       val animSpeed: Float = 1f,
                       val animToEnabledSpeed: Float = 1f)
 interface IBottomClickListener {
-    fun menuClicked(position: Int, id: Long): Boolean
-    fun menuReselected(position: Int, id: Long)
+    fun menuClicked(position: Int, id: Int): Boolean
+    fun menuReselected(position: Int, id: Int)
 }
 
 fun Float.toDp(context: Context): Int {
